requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.4.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
databases==0.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi-slim==0.111.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
PyJWT==2.8.0
python-decouple==3.8
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.1
